backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Authentication
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1

# Docker
docker==6.1.3

# WebSockets
websockets==12.0

# Utilities
python-dotenv==1.0.0
pydantic[email]==2.5.2
pydantic-settings==2.1.0

# Monitoring (optional)
prometheus-client==0.19.0








